Changed Icon Formatting
I learned how to adjust the background image of a CSS property using JS. I was really excited about the usage, but it unfortunately did not turn out well with the images.

I opted to adjust the icon methodology to be images set above the forecast temps in a separate div. It looks much better. CSS/HTML/JS updated.